refactor: modularize dashboard into package, add Typer CLI, move /utils to /scripts
