+++ requirements.txt
@@ -15,3 +15,4 @@ python-dotenv
 tqdm
 Flask
 flask-cors
+typer
