+++ scripts/combine_responses.py
@@ -501,3 +501,4 @@ def main(argv: Optional[List[str]] = None) -> None:
 
 if __name__ == "__main__":
     main()
+
