+++ sycophancy_analysis/cli/__init__.py
@@ -1,0 +1,4 @@
+"""CLI entrypoints for the sycophancy analysis toolkit."""
+from .app import app
+
+__all__ = ['app']
